Stabilize search dialog focus lifecycle
